Cast the values that strict types reject in the newly strict files
A search of the 332 files that this branch makes strict found calls that PHP
used to convert without an error and that now throw a TypeError:

- null passed to an internal function: trim(), strlen(), str_contains(),
  explode(), strtotime(), nl2br(), is_readable() with false.
- An id that an admin form posts as a string, passed to an ?int setter: the
  tax rate titles, the dataflow profile history, the cart rule coupon and
  the tag relations.
- An int passed to a ?string setter: the relation child id of an edited
  order, and the entity value of a rating vote.
- A string passed to a ?float or ?int setter: the default customer group of
  an admin quote, the cart qty of an admin order item, the flat rate cost
  and a forced invoice state.
- The flat rate free boxes, a sum of item quantities, reached an ?int
  accessor. The accessor takes a float now.

Each cast gives the value that the silent conversion gave before. The admin
block and variable name checks passed null to a string parameter, which
already threw before strict types. They now report the validation error.

+++ app/code/core/Mage/Adminhtml/Model/Session/Quote.php
@@ -78,7 +78,7 @@ class Mage_Adminhtml_Model_Session_Quote extends Mage_Core_Model_Session_Abstrac
                 $this->setQuoteId($this->_quote->getId());
             } elseif ($this->getStoreId() && $this->hasCustomerId()) {
                 $this->_quote->setStoreId($this->getStoreId())
-                    ->setCustomerGroupId(Mage::getStoreConfig(self::XML_PATH_DEFAULT_CREATEACCOUNT_GROUP))
+                    ->setCustomerGroupId((int) Mage::getStoreConfig(self::XML_PATH_DEFAULT_CREATEACCOUNT_GROUP))
                     ->assignCustomer($this->getCustomer())
                     ->setIsActive(false)
                     ->save();
